# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DeadJob": {
        "description": "One job that will not be retried, and why (\u00a717.1).",
        "properties": {
          "attempt_count": {
            "title": "Attempt Count",
            "type": "integer"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "job_type": {
            "title": "Job Type",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "requires_human_review": {
            "title": "Requires Human Review",
            "type": "boolean"
          }
        },
        "required": [
          "job_id",
          "job_type",
          "reason",
          "attempt_count",
          "requires_human_review"
        ],
        "title": "DeadJob",
        "type": "object"
      },
...
      "FlagChangeRequest": {
        "additionalProperties": false,
        "description": "Throw or release one switch, and say why.",
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "reason": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "reason"
        ],
        "title": "FlagChangeRequest",
        "type": "object"
      },
      "FlagChangeResponse": {
        "description": "What the switch now says.",
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "set_at": {
            "format": "date-time",
            "title": "Set At",
            "type": "string"
          },
          "set_by": {
            "title": "Set By",
            "type": "string"
          },
          "shadow_mode": {
            "title": "Shadow Mode",
            "type": "boolean"
          },
          "what_happens_next": {
            "title": "What Happens Next",
            "type": "string"
          }
        },
        "required": [
          "key",
          "enabled",
          "reason",
          "set_by",
          "set_at",
          "shadow_mode",
          "what_happens_next"
        ],
        "title": "FlagChangeResponse",
        "type": "object"
      },
      "FlagKey": {
        "description": "The database-backed subset of \u00a717.6.",
        "enum": [
          "global_pause",
          "shadow_mode",
          "outbound_email_disabled",
          "product_disabled",
          "claim_version_disabled"
        ],
        "title": "FlagKey",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationsOverview": {
        "description": "\u00a717.5's dashboard, as far as this repository has data for it.",
        "properties": {
          "candidates_awaiting_review": {
            "title": "Candidates Awaiting Review",
            "type": "integer"
          },
          "claim_invalidations": {
            "title": "Claim Invalidations",
            "type": "integer"
          },
          "dead_job_sample": {
            "items": {
              "$ref": "#/components/schemas/DeadJob"
            },
            "title": "Dead Job Sample",
            "type": "array"
          },
          "dead_jobs": {
            "title": "Dead Jobs",
            "type": "integer"
          },
          "delivery_ambiguous_threads": {
            "title": "Delivery Ambiguous Threads",
            "type": "integer"
          },
          "flags_in_force": {
            "items": {
              "type": "string"
            },
            "title": "Flags In Force",
            "type": "array"
          },
          "jobs_by_state": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Jobs By State",
            "type": "object"
          },
          "not_measured": {
            "items": {
              "type": "string"
            },
            "title": "Not Measured",
            "type": "array"
          },
          "oldest_pending_outbox_age_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Pending Outbox Age Seconds"
          },
          "oldest_queued_job_age_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Queued Job Age Seconds"
          },
          "oldest_review_item_age_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oldest Review Item Age Seconds"
          },
          "outbox_pending": {
            "title": "Outbox Pending",
            "type": "integer"
          },
          "revisions_awaiting_review": {
            "title": "Revisions Awaiting Review",
            "type": "integer"
          },
          "shadow_mode": {
            "title": "Shadow Mode",
            "type": "boolean"
          },
          "suppressed_send_attempts": {
            "title": "Suppressed Send Attempts",
            "type": "integer"
          }
        },
        "required": [
          "shadow_mode",
          "flags_in_force",
          "jobs_by_state",
          "oldest_queued_job_age_seconds",
          "dead_jobs",
          "dead_job_sample",
          "outbox_pending",
          "oldest_pending_outbox_age_seconds",
          "delivery_ambiguous_threads",
          "candidates_awaiting_review",
          "revisions_awaiting_review",
          "oldest_review_item_age_seconds",
          "claim_invalidations",
          "suppressed_send_attempts",
          "not_measured"
        ],
        "title": "OperationsOverview",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Application-owned sales workflow. Shadow mode; live outreach is gated.",
    "title": "Matrix Power Always-On AI Sales Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/operations/flags/{key}": {
      "post": {
        "description": "\u00a717.6's global pause, shadow-mode switch, and outbound-email disable.\n\n`PAUSE_SYSTEM` is tier 5 and administrator-only (\u00a77.4): these are the switches that stop the\nsystem, and \u00a712.1 gives them to nobody else. The actor comes from the session, never from the\nrequest (\u00a715.1, `T-070c`), and `set_flag` writes the audit event.\n\n**A scoped key is refused by `set_flag`, not by a check here.** This route first carried its\nown `key in SCOPED_KEYS` guard, and a negative control proved it dead: deleting it changed\nnothing, because `PRODUCT_DISABLED` with no `scope_id` is already a `FlagError` one layer\ndown. A rule enforced twice is a rule that can disagree with itself, and the copy nobody\nexercises is the one that drifts \u2014 the same reasoning that removed the duplicate\nverification check from `approve_message._recheck`.",
        "operationId": "set_operational_flag_api_operations_flags__key__post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FlagKey"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "mp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mp Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagChangeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Throw or release a system-wide operational switch (administrator only)",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/operations/overview": {
      "get": {
        "description": "Every \u00a717.5 counter this repository has rows for, plus the switches in force.\n\n`principal` is unused in the body and required in the signature: it is what runs the\nauthorization, and naming it keeps that visible at the endpoint rather than hidden in a\ndecorator somebody could remove without the route looking different (`T-063a`'s reasoning).",
        "operationId": "operations_overview_api_operations_overview_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "cookie",
            "name": "mp_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "What the system is doing right now (administrator only)",
        "tags": [
          "operations"
        ]
      }
    },
...
  }
}